+++ x/slashing/keeper/signing_info_test.go
@@ -21,15 +21,17 @@ func TestGetSetValidatorSigningInfo(t *testing.T) {
 	require.False(t, found)
 	newInfo := types.NewValidatorSigningInfo(
 		sdk.ConsAddress(addrDels[0]),
+		int64(4),
+		int64(3),
 		time.Unix(2, 0),
 		false,
 		int64(10),
-		int64(3),
 	)
 	app.SlashingKeeper.SetValidatorSigningInfo(ctx, sdk.ConsAddress(addrDels[0]), newInfo)
 	info, found = app.SlashingKeeper.GetValidatorSigningInfo(ctx, sdk.ConsAddress(addrDels[0]))
 	require.True(t, found)
-	require.Equal(t, info.VoterSetCounter, int64(3))
+	require.Equal(t, info.StartHeight, int64(4))
+	require.Equal(t, info.IndexOffset, int64(3))
 	require.Equal(t, info.JailedUntil, time.Unix(2, 0).UTC())
 	require.Equal(t, info.MissedBlocksCounter, int64(10))
 }
@@ -56,10 +58,11 @@ func TestTombstoned(t *testing.T) {
 
 	newInfo := types.NewValidatorSigningInfo(
 		sdk.ConsAddress(addrDels[0]),
+		int64(4),
+		int64(3),
 		time.Unix(2, 0),
 		false,
 		int64(10),
-		int64(3),
 	)
 	app.SlashingKeeper.SetValidatorSigningInfo(ctx, sdk.ConsAddress(addrDels[0]), newInfo)
 
@@ -78,10 +81,11 @@ func TestJailUntil(t *testing.T) {
 
 	newInfo := types.NewValidatorSigningInfo(
 		sdk.ConsAddress(addrDels[0]),
+		int64(4),
+		int64(3),
 		time.Unix(2, 0),
 		false,
 		int64(10),
-		int64(3),
 	)
 	app.SlashingKeeper.SetValidatorSigningInfo(ctx, sdk.ConsAddress(addrDels[0]), newInfo)
 	app.SlashingKeeper.JailUntil(ctx, sdk.ConsAddress(addrDels[0]), time.Unix(253402300799, 0).UTC())