+++ x/slashing/simulation/operations_test.go
@@ -69,8 +69,8 @@ func TestSimulateMsgUnjail(t *testing.T) {
 	app.StakingKeeper.SetValidatorByConsAddr(ctx, validator0)
 	val0ConsAddress, err := validator0.GetConsAddr()
 	require.NoError(t, err)
-	info := types.NewValidatorSigningInfo(val0ConsAddress,
-		time.Unix(2, 0), false, int64(10), int64(3))
+	info := types.NewValidatorSigningInfo(val0ConsAddress, int64(4), int64(3),
+		time.Unix(2, 0), false, int64(10))
 	app.SlashingKeeper.SetValidatorSigningInfo(ctx, val0ConsAddress, info)
 
 	// put validator0 in jail