+++ x/token/event_test.go
@@ -50,7 +50,8 @@ func TestNewEventIssueToken(t *testing.T) {
 	rand.Seed(time.Now().UnixNano())
 	str := func() string { return randomString(8) }
 
-	event := token.EventIssue{
+	event := token.EventIssued{
+		Creator:    str(),
 		ContractId: str(),
 		Name:       str(),
 		Symbol:     str(),
@@ -59,10 +60,9 @@ func TestNewEventIssueToken(t *testing.T) {
 		Decimals:   0,
 		Mintable:   true,
 	}
-	operator := sdk.AccAddress(str())
 	to := sdk.AccAddress(str())
 	amount := sdk.OneInt()
-	legacy := token.NewEventIssueToken(event, operator, to, amount)
+	legacy := token.NewEventIssueToken(event, to, amount)
 
 	require.Equal(t, token.EventTypeIssueToken.String(), legacy.Type)
 
@@ -75,7 +75,7 @@ func TestNewEventIssueToken(t *testing.T) {
 		token.AttributeKeyMintable:   fmt.Sprintf("%v", event.Mintable),
 		token.AttributeKeyDecimals:   fmt.Sprintf("%d", event.Decimals),
 		token.AttributeKeyAmount:     amount.String(),
-		token.AttributeKeyOwner:      operator.String(),
+		token.AttributeKeyOwner:      event.Creator,
 		token.AttributeKeyTo:         to.String(),
 	}
 	for key, value := range attributes {
@@ -257,7 +257,7 @@ func TestNewEventGrantPermToken(t *testing.T) {
 		return token.Permission(1 + rand.Intn(n))
 	}
 
-	event := token.EventGrant{
+	event := token.EventGranted{
 		ContractId: str(),
 		Granter:    str(),
 		Grantee:    str(),
@@ -286,7 +286,7 @@ func TestNewEventGrantPermTokenHead(t *testing.T) {
 		return token.Permission(1 + rand.Intn(n))
 	}
 
-	event := token.EventGrant{
+	event := token.EventGranted{
 		ContractId: str(),
 		Granter:    str(),
 		Grantee:    str(),
@@ -313,7 +313,7 @@ func TestNewEventGrantPermTokenBody(t *testing.T) {
 		return token.Permission(1 + rand.Intn(n))
 	}
 
-	event := token.EventGrant{
+	event := token.EventGranted{
 		ContractId: str(),
 		Granter:    str(),
 		Grantee:    str(),
@@ -339,7 +339,7 @@ func TestNewEventRevokePermToken(t *testing.T) {
 		return token.Permission(1 + rand.Intn(n))
 	}
 
-	event := token.EventAbandon{
+	event := token.EventRenounced{
 		ContractId: str(),
 		Grantee:    str(),
 		Permission: permission(),