+++ x/wasm/alias.go
@@ -127,7 +127,4 @@ type (
 	CustomQuerier                              = keeper.CustomQuerier
 	QueryPlugins                               = keeper.QueryPlugins
 	Option                                     = keeper.Option
-
-	EncodeHandler = types.EncodeHandler
-	EncodeQuerier = types.EncodeQuerier
 )